src/backend/requirements.txt
certifi==2023.11.17
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-rq==2.10.1
django-taggit==5.0.1
drf-spectacular==0.27.0
pycryptodome==3.19.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-magic==0.4.27
python-libnmap==0.7.3
python-telegram-bot==20.7
pyyaml==6.0.1
requests==2.31.0
rq==1.15.1
stringcase==1.2.0
tornado==6.4